+++ src/CommandLineUtils/Internal/Conventions/IAppConvention.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace McMaster.Extensions.CommandLineUtils.Conventions
-{
-    internal interface IAppConvention
-    {
-        void Apply(CommandLineApplication app, Type modelType);
-    }
-}