+++ src/CommandLineUtils/Internal/Conventions/IConventionBuilder.cs
@@ -1,7 +1,0 @@
-namespace McMaster.Extensions.CommandLineUtils.Conventions
-{
-    internal interface IConventionBuilder
-    {
-        void AddConvention(IAppConvention convention);
-    }
-}