+++ src/CommandLineUtils/Internal/Conventions/IModelProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace McMaster.Extensions.CommandLineUtils.Conventions
-{
-    internal interface IModelProvider
-    {
-        object Model { get; }
-    }
-}